Refactor PkiUtility class
refs #5450
