+++ lib/cli/apisetuputility.cpp
@@ -18,10 +18,10 @@
  ******************************************************************************/
 
 #include "cli/apisetuputility.hpp"
-#include "cli/pkiutility.hpp"
 #include "cli/nodeutility.hpp"
 #include "cli/featureutility.hpp"
 #include "remote/apilistener.hpp"
+#include "remote/pkiutility.hpp"
 #include "base/logger.hpp"
 #include "base/console.hpp"
 #include "base/application.hpp"
