+++ lib/cli/nodesetupcommand.cpp
@@ -20,9 +20,9 @@
 #include "cli/nodesetupcommand.hpp"
 #include "cli/nodeutility.hpp"
 #include "cli/featureutility.hpp"
-#include "cli/pkiutility.hpp"
 #include "cli/apisetuputility.hpp"
 #include "remote/apilistener.hpp"
+#include "remote/pkiutility.hpp"
 #include "base/logger.hpp"
 #include "base/console.hpp"
 #include "base/application.hpp"
