+++ lib/cli/pkinewcertcommand.cpp
@@ -18,7 +18,7 @@
  ******************************************************************************/
 
 #include "cli/pkinewcertcommand.hpp"
-#include "cli/pkiutility.hpp"
+#include "remote/pkiutility.hpp"
 #include "base/logger.hpp"
 
 using namespace icinga;
